Cleaned up CSS and classes that weren't necessary anymore.
Bugfix: Sometimes, seemingly randomly, interests would be hidden when clicking on another filter. I'm not sure where this stems from, but the function to show all cards only set the boolean allIsShown to true and left the others alone. I'm assuming this was affecting the next updateFilters() call. This has been fixed now.